+++ src/lib.rs
@@ -139,15 +139,15 @@ pub mod canvas {
         /// # Panics
         ///
         /// Panics if `m` is negative.
-        pub fn multiply_by(&self, m: f32) -> Rgb {
+        pub fn multiply_by(&self, m: f64) -> Rgb {
             if m < 0.0 {
                 panic!("Rgb cannot be multiplied by a negative value");
             }
 
             Self {
-                red: (self.red as f32 * m) as u8,
-                green: (self.green as f32 * m) as u8,
-                blue: (self.blue as f32 * m) as u8,
+                red: (self.red as f64 * m) as u8,
+                green: (self.green as f64 * m) as u8,
+                blue: (self.blue as f64 * m) as u8,
             }
         }
